==> CDD_training_aids/New_Stuff/Foot B/2_Normalize.Rmd <==
---
title: "Normalize Data"
author: "Wesley Burr"
date: "29/03/2022"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(dplyr)
```

## Normalize Data

We will load each of the data files, identify the 'primary' peak for
Bromobenze, then normalize all Areas to be relative areas to the
reference standard.

```{r}
load("/home/wburr/Rushali_Analysis/New_Stuff/Foot B/FootB.rda")
FootB1m$Area<-as.numeric(FootB1m$Area)
bromo <- FootB1m %>% 
           group_by(Sample) %>% 
           subset(Name == "Benzene, bromo-") %>%
           filter(Area == max(Area))
split_FootB <- FootB1m %>% 
                group_by(Sample) %>%
                group_split()
for(j in 1:length(split_FootB)) {
  split_FootB[[j]]$Area <- split_FootB[[j]]$Area / unlist(bromo[j, "Area"])
}
FootB1m <- do.call("rbind", split_FootB)
save(file = "/home/wburr/Rushali_Analysis/New_Stuff/Foot B/FootBn.rda", FootB1m)
```
